refactor: 약관 동의 부분 구조 변경
