+++ .prettierrc
@@ -2,5 +2,5 @@
   "singleQuote": true,
   "trailingComma": "all",
   "tabWidth": 2,
-  "printWidth": 80
+  "printWidth": 100
 }
