+++ src/components/signup/agreecontent/Agreecontent.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export const AgeCheckContent = () => {
+  return <>14 이상 약관 내용 ~~~~</>;
+};
+export const ServiceCheckContent = () => {
+  return <>서비스 이용 약관 내용 ~~~~</>;
+};
+export const PrivateInfoContent = () => {
+  return <>개인정보 처리방침 약관 내용 ~~~~</>;
+};
+export const AdvertiseCheckContent = () => {
+  return <>마케팅 약관 내용 ~~~~</>;
+};
